+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Replace transistor with MOS-FET in battery operation circuit (Q2) thanks to conditional pullup
 * Self-check if active battery check circuit is populated, not by USE_ACTIVE_BATTERY_CHECK macro definition
 ### Fixed
-* Add workaround for mount fail case of Samsung PRO Plus card when reboot
+* Add workaround for mount fail case of Samsung PRO Plus card
 
 ## [v0.9.4] - 2024-04-09
 ### Added